Add link to LinkedIn in the footer
In a similar way than the twitter and the github are handled.
I had to find a svg for LinkedIn, and use a scale function in
the SVG to make it fit in 16x16 like the others.

As I have a long handler in linkedin, I had to change the width
of the columns in the foot.